refactor: separate changes getter in a file
